Added the autoptr_zero_obj() procedure for zeroing all bytes of the managed object excluding the autoptr struct.
The autoptr_unbind() precedure now calls autoptr_dtor() which requires
a valid autoptr struct during cleanup. This call occurs after the call
to the object destructor.

The autoptr_zero_obj() procedure is to be used in place of memset()
for zeroing the managed object within object destructors in order to
preserve the base autoptr struct.